Add tests for the fingerpostwire
